backend/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
win32-setctime==1.1.0
